service/merchant/requirements.txt
astroid==2.4.1
certifi==2020.4.5.1
chardet==3.0.4
grpcio==1.28.1
grpcio-tools==1.28.1
idna==2.9
isort==4.3.21
JSON-log-formatter==0.3.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
protobuf==3.11.3
psutil==5.7.0
requests==2.23.0
rollbar==0.15.0
six==1.14.0
snowballstemmer==2.0.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.9
wrapt==1.12.1
